requirements.txt
Django>=4.2.6
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
drf-spectacular>=0.26.5
